moved ipynb files to /image_processing_notebook to clean up the repo
