+++ gee_tools.py
@@ -27,20 +27,56 @@ def osm2gee_bbox(bbox):
     return bbox[1], bbox[0], bbox[3], bbox[2]
 
 
-def download_NAIP_toLocal(bbox, name, scale=1):
+def download_sen2_toLocal(bbox, name, scale=30):
     """
     downloads NAIP imagery from the specified bounding box and saves it as `name`
     """
     AOI = ee.Geometry.Rectangle(list(bbox), "EPSG:4326", False)
 
+    start_date = "2020-01-01"
+    end_date = "2020-06-30"
+
     collection = (
-        ee.ImageCollection("USDA/NAIP/DOQQ")
-        .filterDate("2010-01-01", "2019-01-01")
+        ee.ImageCollection("COPERNICUS/S2_SR")
+        .filterDate(start_date, end_date)
+        .filter(ee.Filter.lt('CLOUDY_PIXEL_PERCENTAGE',20))
         .filterBounds(AOI)
+        .select(['B2', 'B3', 'B4'])
     )
 
-    image = ee.Image(collection.mosaic()).clip(AOI)
-    batch.image.toLocal(image, name, scale=scale, region=AOI)
+    # Original code from previous year's:
+    # image = ee.Image(collection.mosaic()).clip(AOI)
+    # batch.image.toLocal(image, name, scale=scale, region=AOI)
+
+    export_image(collection, name, region=AOI)
+
+
+def export_image(collection, folder, scale=10, region=None):
+    """ Batch export images to local directory, one image at a time
+    Arguments
+        :param collection: earth engine image collection object, a stack of images to export
+        :param folder: specify the name of the folder to store images in
+        :param scale=10: resolution of the image. Set to the orginal resolution of Sentinel 2 by default
+        :param region=None: Area of interest to crop the image to. 
+                            If pass in coordinates, the image will be cropped to 
+                            interested range of coordinates
+    """
+
+    colList = collection.toList(collection.size())
+    n = collection.size().getInfo()
+
+    for i in range(n):
+        img = ee.Image(colList.get(i))
+        # imgid = img.id().getInfo()
+        if region is None:
+            region = img.geometry().bounds().getInfo()["coordinates"]
+
+        batch.image.toLocal(
+            image=img,
+            name=folder,
+            region=region,
+            scale=scale
+            )
 
 
 def lc_code_to_str(code):
@@ -76,7 +112,7 @@ if __name__ == "__main__":
         "-d",
         "--distance",
         help="side length for area of interest",
-        default=500,
+        default=1000,
         type=int,
     )
     parser.add_argument(
@@ -149,7 +185,7 @@ if __name__ == "__main__":
             )
             bbox = osm2gee_bbox(bbox)
             try:
-                download_NAIP_toLocal(bbox, fname)
+                download_sen2_toLocal(bbox, fname)
                 os.remove(f"{fname}.zip")
             except Exception as e:
                 logf.write(f"point id {point[args.id_col]}: {e}\n")
